+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/CompletenessGenerator.php
@@ -61,6 +61,4 @@ class CompletenessGenerator implements CompletenessGeneratorInterface
         // @TODO Not implemented yet
         return;
     }
-
-
 }