+++ src/Pim/Bundle/DataGridBundle/Datagrid/Product/FiltersConfigurator.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Pim\Bundle\DataGridBundle\Datagrid\Flexible;
+namespace Pim\Bundle\DataGridBundle\Datagrid\Product;
 
 use Oro\Bundle\DataGridBundle\Datagrid\Common\DatagridConfiguration;
 use Oro\Bundle\FilterBundle\Grid\Extension\Configuration as FilterConfiguration;
 use Pim\Bundle\FilterBundle\Filter\Flexible\FilterUtility;
-use Pim\Bundle\DataGridBundle\Datasource\Orm\OrmDatasource;
+use Pim\Bundle\DataGridBundle\Datasource\ProductDatasource;
 
 /**
  * Filters configurator for flexible grid
@@ -51,7 +51,7 @@ class FiltersConfigurator implements ConfiguratorInterface
      */
     public function configure()
     {
-        $attributes = $this->configuration->offsetGetByPath(OrmDatasource::USEABLE_ATTRIBUTES_PATH);
+        $attributes = $this->configuration->offsetGetByPath(ProductDatasource::USEABLE_ATTRIBUTES_PATH);
 
         foreach ($attributes as $attributeCode => $attribute) {
             $showFilter        = $attribute['useableAsGridFilter'];