+++ CHANGELOG.md
@@ -51,7 +51,7 @@ technical flaws (attribute codes are data, not a freeze structure, needed to mai
 that made the entity too smart for its own good and created performances problem)
 - Remove Product::createValue(). Can be replaced by calling ProductManager::createFlexibleValue() and setting attribute, scope and locale on the created value.
 - Product datagrid, hydrate rows as arrays (in place of objects) to reduce the loading time
-- Datagrid configuration, replace [flexible_entity] config by a [source][is_flexible] config to avoid to define the used entity twice
+- Datagrid configuration, remove [flexible_entity] config to avoid to define the used entity twice
 - Rename and move src/Pim/Bundle/EnrichBundle/Resources/views/Completeness/_datagridCompleteness.html.twig => DataGridBundle/Resources/views/Property/completeness.html.twig
 - Delete classes ConfigureAssociationProductGridListener and AssociationProductColumnsConfigurator, we now use ConfigureFlexibleGridListener to configure product association grid
 - Delete the HideColumnsListener, the ColumnConfigurator is now able to add only columns configured by the user
@@ -61,6 +61,11 @@ that made the entity too smart for its own good and created performances problem
 - ProductManager and FlexibleManager constructors signatures have changed and takes now an array with flexibleConfig as first parameters (see CatalogBundle managers.yml file for more details)
 - CompletenessGeneratorInterface has a new method schedule to schedule completeness of a product
 - Move LocalizableListener and ScopableListener from FlexibleEntityBundle to CatalogBundle and make them product specific
+- Introduce a pim_product datasource to prepare the MongoDB support
+- Move DatagridBundle/Datagrid/Flexible to DatagridBundle/Datagrid/Product
+- Change configuration key in DataGridBundle/Resources/config/grid_attribute_types.yml
+- rename ConfigureFlexibleGridListener to ConfigureProductGridListener
+- Change Datagrid/Product/SortersConfigurator constructor (remove FlexibleManager argument)
 
 # 1.0.0-rc-1 - "Tortoise Beats Hare" (2014-02-06)